requirements.txt
pampy ~= 0.3.0
PyYAML ~= 5.4.1
termcolor ~= 1.1.0
websockets ~= 8.1
pytest ~= 6.2.2
pytest-asyncio ~= 0.14.0
pyhumps ~= 1.6.1
tinydb ~= 4.4.0
